Hacky initial implementation for text fields
Proof of concept.

Since we can't make Ranges for the content of a field or use
getClientBoundingRect, the out-of-box test now compares the current
range to the one when the box popped-up to know when we've moved.
Currently it now changes every time you move one character instead
of only when you move a whole word. This could be changed by
remembering the length of the current match and checking if the
offset is within lastOffset + this length.

Hard-coded pixel values need to go too.